complete trip cycle and stop at start trip

+++ lib/features/RideFeature/domain/entities/request_trip_params.dart
@@ -48,7 +48,7 @@ class RequestTripUseCaseParams {
     "price": price,
     "fromTitle": fromTitle,
     "toTitle": toTitle,
-    "distance": distance,
+    "distance": distance.toInt(),
     "duration": duration,
     "startLocation": startLocation,
     "targetLocation": targetLocation,
